integrate FalconMamba inside mamba.py
