+++ vllm/model_executor/layers/layernorm.py
@@ -14,24 +14,19 @@ class RMSNorm(CustomOp):
     Refer to https://arxiv.org/abs/1910.07467
     """
 
-    def __init__(self,
-                 hidden_size: int,
-                 eps: float = 1e-6,
-                 var_hidden_size: Optional[int] = None,
-                 is_learnable: bool = True) -> None:
+    def __init__(
+        self,
+        hidden_size: int,
+        eps: float = 1e-6,
+        var_hidden_size: Optional[int] = None,
+    ) -> None:
         super().__init__()
 
         self.hidden_size = hidden_size
         self.variance_epsilon = eps
         self.variance_size_override = (None if var_hidden_size == hidden_size
                                        else var_hidden_size)
-        if is_learnable:
-            self.register_parameter("weight",
-                                    nn.Parameter(torch.ones(hidden_size)))
-        else:
-            self.register_buffer('weight',
-                                 torch.ones(hidden_size),
-                                 persistent=False)
+        self.weight = nn.Parameter(torch.ones(hidden_size))
 
     def forward_native(
         self,
